+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,12 @@
 ﻿# @addfox/utils
 
+## 0.2.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @addfox/common@0.2.6
+
 ## 0.2.5
 
 ### Patch Changes